fix: split protocol selector for used in enterprise pages
